+++ app/Http/Controllers/MessageController.php
@@ -22,18 +22,18 @@ class MessageController extends Controller
                 'user_id' => 'required|integer',
                 'text' => 'required|string',
             ]);
-    
+
             if ($validator->fails()) {
                 return response()->json([
                     'Validation failed',
                     $validator->errors()
                 ], 400);
             }
-    
+
             $message = Message::create([
                 'party_id' => $request->party_id,
                 'user_id' => $request->user_id,
-                'text'=> $request->text
+                'text' => $request->text
             ]);
 
             $data = [
@@ -112,26 +112,37 @@ class MessageController extends Controller
 
         try {
 
+            //Get user by auth token
+            $userAuth = auth()->user();
+
             $validator = Validator::make($request->all(), [
                 'text' => 'required|string',
-            
+
             ]);
-    
+
             if ($validator->fails()) {
                 return response()->json([
                     'Validation failed',
                     $validator->errors()
                 ], 400);
             }
-    
+
             $message = Message::find($id);
-    
+
             if (!$message) {
                 return response()->json([
                     'message' => 'Message not found'
                 ], 404);
             }
-    
+
+            //The message can only be updated by an Admin or the user who created it
+            if (($userAuth->isAdmin == false) && ($userAuth->id != $message->user_id)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => "You don't have permissions to perform this action"
+                ], 400);
+            }
+
             $message->text = $request->text;
 
             $data = [
@@ -154,6 +165,9 @@ class MessageController extends Controller
 
         try {
 
+            //Get user by auth token
+            $userAuth = auth()->user();
+
             $message = Message::find($id);
 
             if (!$message) {
@@ -161,14 +175,20 @@ class MessageController extends Controller
                     'message' => 'Message not found'
                 ], 404);
             }
-    
+
+            //The message can only be deleted by an Admin or the user who created it
+            if (($userAuth->isAdmin == false) && ($userAuth->id != $message->user_id)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => "You don't have permissions to perform this action"
+                ], 400);
+            }
+
             $message->delete();
-    
+
             return response()->json([
                 'message' => 'Message deleted'
             ], 200);
-
-
         } catch (\Exception $e) {
 
             return response()->json([
@@ -177,5 +197,4 @@ class MessageController extends Controller
             ], 400);
         }
     }
-
 }
